Fix warning about flags for setFrameStyle()

+++ widgets/settings/settings.h
@@ -55,7 +55,7 @@ struct SYNCTHINGWIDGETS_EXPORT NotifyOn {
 struct SYNCTHINGWIDGETS_EXPORT Appearance {
     bool showTraffic = true;
     QSize trayMenuSize = QSize(450, 400);
-    int frameStyle = QFrame::NoFrame | QFrame::Plain;
+    int frameStyle = static_cast<int>(QFrame::NoFrame) | static_cast<int>(QFrame::Plain);
     int tabPosition = QTabWidget::South;
     bool brightTextColors = false;
     struct SYNCTHINGWIDGETS_EXPORT Positioning {
